mcp-n8n-proxy/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-n8n-proxy"
version = "0.1.0"
description = "MCP server for controlling n8n workflows via MCP tools"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "mcp>=0.9.0",
    "httpx>=0.27.0",
    "python-dotenv>=1.0.0",
    "fastmcp>=2.13.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
http = [
    "uvicorn>=0.27.0",
    "starlette>=0.36.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["mcp_n8n_proxy*"]
